(feat) Added react navigator
